+++ shop-ui-admin/src/main/java/ru/abtank/servise/ProductTypeService.java
@@ -10,9 +10,9 @@ public interface ProductTypeService {
 
     List<ProductTypeRepr> findAll();
 
-    Optional<ProductTypeRepr> findById(Long id);
+    Optional<ProductTypeRepr> findById(Integer id);
 
-    void deleteById(Long id);
+    void deleteById(Integer id);
 
     void save(ProductTypeRepr productTypeRepr) throws IOException;
 }
