+++ shop-ui-admin/src/main/java/ru/abtank/servise/StatusService.java
@@ -10,9 +10,9 @@ public interface StatusService {
 
     List<StatusRepr> findAll();
 
-    Optional<StatusRepr> findById(Long id);
+    Optional<StatusRepr> findById(Integer id);
 
-    void deleteById(Long id);
+    void deleteById(Integer id);
 
     void save(StatusRepr statusRepr) throws IOException;
 }
